Move `getExpectedAtiDestination` out of Cypress into its own file
