+++ src/app/components/ATIAnalytics/amp/ampAnalyticsJson.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-template-curly-in-string */
-import { getExpectedAtiDestination } from '#nextjs/cypress/e2e/specialFeatures/atiAnalytics/helpers';
+import { getExpectedAtiDestination } from '../helpers/getExpectedAtiDestination';
 import { ReverbBeaconConfig } from '../types';
 import ampAnalyticsJson from './ampAnalyticsJson';
